separating interfaces with classes (task #6433)
